% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{coef,ubmsFit-method}
\alias{coef,ubmsFit-method}
\title{Extract Coefficient Values From a ubmsFit Model}
\usage{
\S4method{coef}{ubmsFit}(object, ...)
}
\arguments{
\item{object}{A \code{ubmsFit} model}

\item{...}{Currently ignored}
}
\value{
A vector of coefficient values for all submodels.
}
\description{
Extract Coefficient Values From a ubmsFit Model
}
