% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{default_priors}
\alias{default_priors}
\title{Default priors for single-season models}
\usage{
default_priors()
}
\value{
A list of lists of default prior specifications, one entry
 per submodel.
}
\description{
Convenience function providing default priors for most \code{ubms} models.
Also useful for seeing correct list structure for custom priors.
}
\examples{
default_priors()

}
