% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{extract,ubmsFit-method}
\alias{extract,ubmsFit-method}
\title{Extract Samples From a ubmsFit Model}
\usage{
\S4method{extract}{ubmsFit}(object, pars, permuted = TRUE, inc_warmup = FALSE, include = TRUE)
}
\arguments{
\item{object}{A \code{ubmsFit} object}

\item{pars}{An optional character vector providing parameter
names of interest. If not specified, all parameters are used}

\item{permuted}{Logical. If \code{TRUE}, draws are permuted
and merged; if \code{FALSE}, the original order is kept}

\item{inc_warmup}{Logical. If \code{TRUE}, warmup iterations
are included; if \code{FALSE} they are discarded.}

\item{include}{Logical. If \code{TRUE} provided parameter names
in \code{pars} are kept; if \code{FALSE} they are excluded.}
}
\value{
If \code{permuted=TRUE}, a list; if \code{permuted=FALSE},
 an array.
}
\description{
Extract samples from a \code{ubmsFit} model
}
