% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlist.R
\name{fitList,ubmsFit-method}
\alias{fitList,ubmsFit-method}
\alias{fitList}
\title{Create a List of ubmsFit Models}
\usage{
\S4method{fitList}{ubmsFit}(...)
}
\arguments{
\item{...}{\code{ubmsFit} model objects, preferably named}
}
\value{
An object of class \code{ubmsFitList} containing the list of models
}
\description{
Create a list of ubmsFit models
}
