% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{get_elapsed_time,ubmsFit-method}
\alias{get_elapsed_time,ubmsFit-method}
\title{Get Model Runtime}
\usage{
\S4method{get_elapsed_time}{ubmsFit}(object, ...)
}
\arguments{
\item{object}{A \code{ubmsFit} object}

\item{...}{Arguments passed to \code{rstan::get_elapsed_time}}
}
\value{
A matrix with one row per chain and two columns, containing
 the warmup time and sampling time, respectively, in seconds
}
\description{
Get warmup and sampling time from a \code{ubmsFit} object
}
