% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{get_stancode,ubmsFit-method}
\alias{get_stancode,ubmsFit-method}
\title{Get Stan Code From Model}
\usage{
\S4method{get_stancode}{ubmsFit}(object, ...)
}
\arguments{
\item{object}{A \code{ubmsFit} object}

\item{...}{Arguments passed to \code{rstan::get_stancode}}
}
\value{
A character string with the model code
}
\description{
Get the Stan code used to run a model as a character string
}
\details{
Pass the result of \code{get_stancode} to \code{cat} to get the
 code in a more readable format. Note that the output in most cases
 is Stan code that can be used to fit several types of models, and not
 all Stan code will be used in all models.
}
