% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{getY,ubmsFit-method}
\alias{getY,ubmsFit-method}
\title{Extract y, the Response Variable, From a ubmsFit Model}
\usage{
\S4method{getY}{ubmsFit}(object)
}
\arguments{
\item{object}{A \code{ubmsFit} model}
}
\value{
A matrix containing the response variable \code{y}.
}
\description{
Extract y, the Response Variable, From a ubmsFit Model
}
