% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\docType{methods}
\name{gof}
\alias{gof}
\alias{gof,ubmsFitOccu-method}
\title{Check model goodness-of-fit}
\usage{
gof(object, draws = NULL, ...)

\S4method{gof}{ubmsFitOccu}(object, draws = NULL, ...)
}
\arguments{
\item{object}{A fitted model of class \code{ubmsFit}}

\item{draws}{Number of draws from the posterior to use in the check}

\item{...}{Currently ignored}
}
\value{
An object of class \code{ubmsGOF} containing statistics calculated
from the posterior predictive distribution.
}
\description{
Goodness-of-fit tests for \code{ubmsFit} models using posterior predictive
checks
}
\section{Methods (by class)}{
\itemize{
\item \code{ubmsFitOccu}: Applies the MacKenzie-Bailey chi-square goodness of fit test for
ocupancy models (MacKenzie and Bailey 2004).
}}

\references{
MacKenzie, D. I., & Bailey, L. L. (2004). Assessing the 
 fit of site-occupancy models. Journal of Agricultural, Biological, 
 and Environmental Statistics, 9(3), 300-318.
}
