% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlist.R
\name{modSel,ubmsFitList-method}
\alias{modSel,ubmsFitList-method}
\title{Model Selection For a List of ubmsFit Models}
\usage{
\S4method{modSel}{ubmsFitList}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ubmsFitList}}

\item{...}{Currently ignored}
}
\value{
A \code{data.frame} of model fit information with one row per
 model in the input \code{ubmsFitList}. Models are ranked in descending
 order by expected log pointwise predictive density (\code{elpd}).
}
\description{
Construct a model selection table from a \code{ubmsFitList}
}
\seealso{
\code{\link[loo]{loo}}, \code{\link[loo]{loo_compare}}
}
