% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{names,ubmsFit-method}
\alias{names,ubmsFit-method}
\title{Get Parameter Names From a ubmsFit Model}
\usage{
\S4method{names}{ubmsFit}(x)
}
\arguments{
\item{x}{A \code{ubmsFit} model}
}
\value{
A character vector of parameter names.
}
\description{
Get Parameter Names From a ubmsFit Model
}
