% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFitList-methods.R
\name{names,ubmsFitList-method}
\alias{names,ubmsFitList-method}
\title{Get Names of Models in a ubmsFitList}
\usage{
\S4method{names}{ubmsFitList}(x)
}
\arguments{
\item{x}{A \code{ubmsFitList} object}
}
\value{
A character vector of model names.
}
\description{
Get Names of Models in a ubmsFitList
}
