% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{normal}
\alias{normal}
\title{Specify normal prior distribution}
\usage{
normal(location = 0, scale = 10, autoscale = TRUE)
}
\arguments{
\item{location}{The mean of the distribution}

\item{scale}{The standard deviation of the distribution}

\item{autoscale}{If \code{TRUE}, ubms will automatically adjust priors
for each regression coefficient relative to its corresponding covariate x.
Specifically, the prior for a given coefficient will be divided by
sd(x). This helps account for covariates with very different magnitudes
in the same model. If your data are already standardized (e.g. with use of
\code{scale()}), this will have minimal effect as sd(x) will be
approximately 1. Standardizing your covariates is highly recommended.}
}
\value{
A \code{list} containing prior settings used internally by \code{ubms}.
}
\description{
Controls settings for normal prior distributions on an intercept or
on coefficients for a submodel.
}
\examples{
normal()

}
