% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{nsamples,ubmsFit-method}
\alias{nsamples,ubmsFit-method}
\title{Get number of Posterior Samples Stored in a ubmsFit Model}
\usage{
\S4method{nsamples}{ubmsFit}(object, ...)
}
\arguments{
\item{object}{A \code{ubmsFit} model}

\item{...}{Currently ignored}
}
\value{
An integer representing the number of posterior samples
}
\description{
Get number of Posterior Samples Stored in a ubmsFit Model
}
