% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{plot,ubmsFit,ANY-method}
\alias{plot,ubmsFit,ANY-method}
\title{Plot Residuals For All Submodels in a ubmsFit Model}
\usage{
\S4method{plot}{ubmsFit,ANY}(x, y, ...)
}
\arguments{
\item{x}{A \code{ubmsFit} model}

\item{y}{Currently ignored}

\item{...}{Currently ignored}
}
\value{
A plot object of class \code{gtable} with one panel per submodel.
}
\description{
Plot Residuals For All Submodels in a ubmsFit Model
}
