% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot_spatial}
\alias{plot_spatial}
\title{Plot A Map of the State Parameter Based on a Spatial ubms Model}
\usage{
plot_spatial(object, param = c("state", "eta"), sites = TRUE)
}
\arguments{
\item{object}{A \code{ubmsFit} model with a spatial random effect}

\item{param}{The parameter to plot, either \code{"state"} for, e.g.,
mean occupancy or abundance, or \code{"eta"} for the random effect itself}

\item{sites}{If \code{TRUE}, also plot the locations of sites that
were sampled on the map and if had a detection of the species}
}
\description{
Plot A Map of the State Parameter Based on a Spatial ubms Model
}
