% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distamp.R
\name{stan_distsamp}
\alias{stan_distsamp}
\title{Fit the Royle et al. (2004) Distance Sampling Model}
\usage{
stan_distsamp(
  formula,
  data,
  keyfun = c("halfnorm", "exp"),
  output = c("density", "abund"),
  unitsOut = c("ha", "kmsq"),
  ...
)
}
\arguments{
\item{formula}{Double right-hand side formula describing covariates of
detection and occupancy in that order}

\item{data}{A \code{\link{unmarkedFrameDS}} object}

\item{keyfun}{One of the following detection functions:
\code{"halfnorm"} for half-normal or \code{"exp"} for negative exponential}

\item{output}{Model either density \code{"density"} or abundance \code{"abund"}}

\item{unitsOut}{Units of density. Either \code{"ha"} or \code{"kmsq"} for
hectares and square kilometers, respectively.}

\item{...}{Arguments passed to the \code{\link{stan}} call, such as
number of chains \code{chains} or iterations \code{iter}}
}
\value{
\code{ubmsFitDistsamp} object describing the model fit.
}
\description{
This function fits the hierarchical distance sampling model of Royle
et al. (2004) to line or point transect data recorded in discerete
distance intervals.
}
\references{
Royle, J. A., Dawson, D. K., & Bates, S. (2004). Modeling
 abundance effects in distance sampling. Ecology 85: 1591-1597.
}
\seealso{
\code{\link{distsamp}}, \code{\link{unmarkedFrameDS}}
}
