% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu.R
\name{stan_occu}
\alias{stan_occu}
\title{Fit the MacKenzie et al. (2002) Occupancy Model}
\usage{
stan_occu(formula, data, ...)
}
\arguments{
\item{formula}{Double right-hand side formula describing covariates of
detection and occupancy in that order}

\item{data}{A \code{\link{unmarkedFrameOccu}} object}

\item{...}{Arguments passed to the \code{\link{stan}} call, such as 
number of chains \code{chains} or iterations \code{iter}}
}
\value{
\code{ubmsFitOccu} object describing the model fit.
}
\description{
This function fits the single season occupancy model of
MacKenzie et al. (2002).
}
\seealso{
\code{\link{occu}}, \code{\link{unmarkedFrameOccu}}
}
