% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcount.R
\name{stan_pcount}
\alias{stan_pcount}
\title{Fit the N-mixture model of Royle (2004)}
\usage{
stan_pcount(formula, data, K = NULL, mixture = "P", ...)
}
\arguments{
\item{formula}{Double right-hand side formula describing covariates of
detection and abundance in that order}

\item{data}{A \code{\link{unmarkedFramePCount}} object}

\item{K}{Integer upper index of integration for N-mixture. This should be
set high enough so that it does not affect the parameter estimates.
Note that computation time will increase with K.}

\item{mixture}{Character specifying mixture: "P" is only option currently.}

\item{...}{Arguments passed to the \code{\link{stan}} call, such as
number of chains \code{chains} or iterations \code{iter}}
}
\value{
\code{ubmsFitPcount} object describing the model fit.
}
\description{
This function fits the single season N-mixture model of
Royle et al. (2002).
}
\seealso{
\code{\link{pcount}}, \code{\link{unmarkedFramePCount}}
}
