% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{[,ubmsFit,character,missing,missing-method}
\alias{[,ubmsFit,character,missing,missing-method}
\title{Extract a Submodel from a ubmsFit Model}
\usage{
\S4method{[}{ubmsFit,character,missing,missing}(x, i)
}
\arguments{
\item{x}{A \code{ubmsFit} model}

\item{i}{The name of a submodel to extract}
}
\value{
An object of class \code{ubmsSubmodel}.
}
\description{
Extract a Submodel from a ubmsFit Model
}
