% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel.R
\name{[,ubmsSubmodelList,character,missing,missing-method}
\alias{[,ubmsSubmodelList,character,missing,missing-method}
\title{Extract a ubmsSubmodel From a ubmsSubmodelList Object}
\usage{
\S4method{[}{ubmsSubmodelList,character,missing,missing}(x, i)
}
\arguments{
\item{x}{Object of class \code{ubmsSubmodelList}}

\item{i}{The name of a submodel}
}
\value{
An object of class \code{ubmsSubmodel}.
}
\description{
Extract a ubmsSubmodel From a ubmsSubmodelList Object
}
