% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{traceplot,ubmsFit-method}
\alias{traceplot,ubmsFit-method}
\title{Markov Chain Traceplots}
\usage{
\S4method{traceplot}{ubmsFit}(object, ...)
}
\arguments{
\item{object}{A \code{ubmsFit} object}

\item{...}{Arguments passed to \code{rstan::traceplot}}
}
\value{
A \code{ggplot} object.
}
\description{
Draws traceplots for chains from a \code{ubmsFit} object
}
