% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colext.R
\name{turnover}
\alias{turnover}
\alias{turnover,ubmsFitColext-method}
\title{Turnover Probability}
\usage{
turnover(object, ...)

\S4method{turnover}{ubmsFitColext}(object, draws, re.form = NULL, ...)
}
\arguments{
\item{object}{A fitted dynamic occupancy model of class inheriting \code{ubmsFit}}

\item{...}{Currently ignored}

\item{draws}{Number of draws from the posterior to use in the check}

\item{re.form}{If \code{NULL}, any estimated group-level parameters ("random
effects") are included. If \code{NA}, they are ignored}
}
\value{
A matrix of turnover values from the posterior predictive distribution.
  The dimensions are \code{draws} by number of sites x (primary periods - 1)
  in site-major order.
}
\description{
Generate posterior draws of turnover probability from dynamic occupancy models.
Turnover is calculated for each site and each primary period after the first.
}
\seealso{
\code{\link{stan_colext}}
}
