% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubms-package.R
\docType{package}
\name{ubms}
\alias{ubms}
\alias{ubms-package}
\title{ubms}
\description{
Unmarked Bayesian Models using Stan
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kenkellner.com/ubms/}
  \item Report bugs at \url{https://github.com/kenkellner/ubms/issues}
}

}
\author{
Ken Kellner
}
