% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFitList-methods.R
\name{ubmsFitList-extractors}
\alias{ubmsFitList-extractors}
\alias{$,ubmsFitList-method}
\alias{[[,ubmsFitList,numeric,missing-method}
\alias{[,ubmsFitList,numeric,missing,missing-method}
\title{Extractors for ubmsFitList objects
Extract parts of ubmsFitList objects.}
\usage{
\S4method{$}{ubmsFitList}(x, name)

\S4method{[[}{ubmsFitList,numeric,missing}(x, i)

\S4method{[}{ubmsFitList,numeric,missing,missing}(x, i)
}
\arguments{
\item{x}{A list of \code{ubmsFit} models of class \code{ubmsFitList}}

\item{name, i}{The names or indices of \code{ubmsFit} models in the \code{ubmsFitList}}
}
\value{
A \code{ubmsFit} object or list of such objects.
}
\description{
Extractors for ubmsFitList objects
Extract parts of ubmsFitList objects.
}
