% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFitList-methods.R
\name{ubmsFitList-methods}
\alias{ubmsFitList-methods}
\alias{$,ubmsFitList-method}
\alias{names,ubmsFitList-method}
\alias{[[,ubmsFitList,numeric,missing-method}
\alias{[,ubmsFitList,numeric,missing,missing-method}
\title{Methods for ubmsFitList objects
Basic methods for ubmsFitList objects. More complex methods have
their own documentation pages.}
\usage{
\S4method{$}{ubmsFitList}(x, name)

\S4method{names}{ubmsFitList}(x)

\S4method{[[}{ubmsFitList,numeric,missing}(x, i)

\S4method{[}{ubmsFitList,numeric,missing,missing}(x, i)
}
\arguments{
\item{x}{A list of \code{ubmsFit} models of class \code{ubmsFitList}}

\item{name, i}{A single \code{ubmsFit} model}
}
\description{
Methods for ubmsFitList objects
Basic methods for ubmsFitList objects. More complex methods have
their own documentation pages.
}
