% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel.R
\name{ubmsSubmodelList-methods}
\alias{ubmsSubmodelList-methods}
\alias{[,ubmsSubmodelList,character,missing,missing-method}
\title{Methods for ubmsSubmodelList objects}
\usage{
\S4method{[}{ubmsSubmodelList,character,missing,missing}(x, i)
}
\arguments{
\item{x}{Object of class \code{ubmsSubmodelList}}

\item{i}{The name of a submodel}
}
\description{
Basic methods for ubmsSubmodelList objects
}
