% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{waic,ubmsFit-method}
\alias{waic,ubmsFit-method}
\title{Widely Applicable Information Criterion (WAIC)}
\usage{
\S4method{waic}{ubmsFit}(x, ...)
}
\arguments{
\item{x}{A \code{ubmsFit} model}

\item{...}{Currently ignored}
}
\value{
An object of class \code{waic} containing an estimate of WAIC and
 other parameters useful for model comparison. See \code{?loo::waic} for
 more information.
}
\description{
Widely Applicable Information Criterion (WAIC)
}
