% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcount.R
\name{stan_pcount}
\alias{stan_pcount}
\title{Fit the N-mixture model of Royle (2004)}
\usage{
stan_pcount(
  formula,
  data,
  K = NULL,
  mixture = "P",
  prior_intercept_state = normal(0, 5),
  prior_coef_state = normal(0, 2.5),
  prior_intercept_det = logistic(0, 1),
  prior_coef_det = logistic(0, 1),
  prior_sigma = gamma(1, 1),
  log_lik = TRUE,
  ...
)
}
\arguments{
\item{formula}{Double right-hand side formula describing covariates of
detection and abundance in that order}

\item{data}{A \code{\link{unmarkedFramePCount}} object}

\item{K}{Integer upper index of integration for N-mixture. This should be
set high enough so that it does not affect the parameter estimates.
Note that computation time will increase with K.}

\item{mixture}{Character specifying mixture: "P" is only option currently.}

\item{prior_intercept_state}{Prior distribution for the intercept of the
state (abundance) model; see \code{?priors} for options}

\item{prior_coef_state}{Prior distribution for the regression coefficients of
the state model}

\item{prior_intercept_det}{Prior distribution for the intercept of the
detection probability model}

\item{prior_coef_det}{Prior distribution for the regression coefficients of
the detection model}

\item{prior_sigma}{Prior distribution on random effect standard deviations}

\item{log_lik}{If \code{TRUE}, Stan will save pointwise log-likelihood values
in the output. This can greatly increase the size of the model. If
\code{FALSE}, the values are calculated post-hoc from the posteriors}

\item{...}{Arguments passed to the \code{\link{stan}} call, such as
number of chains \code{chains} or iterations \code{iter}}
}
\value{
\code{ubmsFitPcount} object describing the model fit.
}
\description{
This function fits the single season N-mixture model of
Royle et al. (2004).
}
\examples{
\donttest{
data(mallard)
mallardUMF <- unmarkedFramePCount(mallard.y, siteCovs=mallard.site)

(fm_mallard <- stan_pcount(~1~elev+forest, mallardUMF, K=30,
                           chains=3, iter=300))
}

}
\references{
Royle JA. 2004. N-mixture models for estimating populaiton size
 from spatially replicated counts. Biometrics 60: 105-108.
}
\seealso{
\code{\link{pcount}}, \code{\link{unmarkedFramePCount}}
}
