% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\name{summary,ubmsFit-method}
\alias{summary,ubmsFit-method}
\title{Extract Summary Statistics from a ubmsFit Model}
\usage{
\S4method{summary}{ubmsFit}(object, submodel, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\arguments{
\item{object}{A \code{ubmsFit} model}

\item{submodel}{Name of submodel to summarize}

\item{probs}{Numeric vector of quantiles of interest}

\item{...}{Currently ignored}
}
\value{
An object of class \code{data.frame} containing summary statistics
 for posterior distributions of parameters from the chosen submodel.
}
\description{
Extract Summary Statistics from a ubmsFit Model
}
