% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmsFit-methods.R
\docType{methods}
\name{ubmsFit-methods}
\alias{ubmsFit-methods}
\alias{[,ubmsFit,character,missing,missing-method}
\alias{nsamples,ubmsFit-method}
\alias{summary,ubmsFit-method}
\alias{loo,ubmsFit-method}
\alias{waic,ubmsFit-method}
\title{Methods for ubmsFit objects
These methods are similar to methods defined for objects of class \code{lm},
\code{merMod}, \code{stanreg}. The most important methods are documented
separately. Links to those pages are provided in the \strong{See Also}
section below.}
\usage{
\S4method{[}{ubmsFit,character,missing,missing}(x, i)

\S4method{nsamples}{ubmsFit}(object, ...)

\S4method{summary}{ubmsFit}(object, submodel, ...)

\S4method{loo}{ubmsFit}(x, ..., cores = getOption("mc.cores", 1))

\S4method{waic}{ubmsFit}(x, ...)
}
\arguments{
\item{object, x}{A fitted model of class \code{ubmsFit}}

\item{...}{Currently ignored}

\item{submodel, i}{A submodel, e.g. \code{"det"} for the detection model}
}
\description{
Methods for ubmsFit objects
These methods are similar to methods defined for objects of class \code{lm},
\code{merMod}, \code{stanreg}. The most important methods are documented
separately. Links to those pages are provided in the \strong{See Also}
section below.
}
\seealso{
\itemize{
 \item The \code{\link{posterior_predict}} for drawing from the posterior
   distribution of the observed outcome or unobserved latent state
}
}
